refactor: C++スタイルのキャストに変更

+++ CPP_Module_02/ex01/srcs/Fixed.cpp
@@ -6,7 +6,7 @@
 /*   By: mizusato <mizusato@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/20 01:14:18 by mizusato          #+#    #+#             */
-/*   Updated: 2025/10/20 01:14:19 by mizusato         ###   ########.fr       */
+/*   Updated: 2025/10/26 14:41:56 by mizusato         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -77,7 +77,7 @@ void	Fixed::setRawBits(const int raw)
 // 固定小数点数からfloatへの変換
 float	Fixed::toFloat(void) const
 {
-	return ((float)_fixedPointValue / (float)(1 << _fractionalBits));
+	return (static_cast<float>(_fixedPointValue) / (1 << _fractionalBits));
 }
 
 // 固定小数点数からintへの変換
